extract out skills combobox
