+++ src/actions/skills.cron.ts
@@ -5,25 +5,16 @@ import { bearerToken } from '@/config/skillapi.auth.token';
 
 var cron = require('node-cron');
 
-let isCronJobRunning = false;
 const url = 'https://auth.emsicloud.com/connect/token';
 
 export async function startAuthTokenCronJob() {
   // don't run the cron job if its already running
-  if (!isCronJobRunning) {
-    isCronJobRunning = true;
-    fetchAuthTokenCronJob();
-    // Schedule the cron to run every 45 minutes , the token resets after one hour
-    cron.schedule('*/45 * * * *', async () => {
-      await fetchAuthTokenCronJob();
-    });
-
-    console.log('Auth token cron job started.');
-    return NextResponse.json({ data: 'Success', status: 200 });
-  } else {
-    console.log('Auth token cron job is already running.');
-    return NextResponse.json({ data: 'Already running', status: 200 });
-  }
+  fetchAuthTokenCronJob();
+  // Schedule the cron to run every 45 minutes , the token resets after one hour
+  cron.schedule('*/10 * * * * *', async () => {
+    await fetchAuthTokenCronJob();
+  });
+  return NextResponse.json({ data: 'Success', status: 200 });
 }
 
 const options = {
