+++ src/components/job-form.tsx
@@ -1,5 +1,4 @@
 'use client';
-import { useDebounce } from '@uidotdev/usehooks';
 import { createJob } from '@/actions/job.action';
 import {
   Form,
@@ -26,7 +25,7 @@ import {
 import { Button } from './ui/button';
 import { Input } from './ui/input';
 import { useToast } from './ui/use-toast';
-import { Calendar, LucideRocket, MailOpenIcon, X } from 'lucide-react';
+import { Calendar, LucideRocket, MailOpenIcon } from 'lucide-react';
 import DescriptionEditor from './DescriptionEditor';
 import Image from 'next/image';
 import { FaFileUpload } from 'react-icons/fa';
@@ -43,8 +42,7 @@ import _ from 'lodash';
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
 import APP_PATHS from '@/config/path.config';
-import { Combobox } from './comboBox';
-import { getBearerToken } from '@/actions/skills.cron';
+import { SkillsCombobox } from './skills-combobox';
 
 const PostJobForm = () => {
   const session = useSession();
@@ -167,57 +165,9 @@ const PostJobForm = () => {
   const watchHasSalaryRange = form.watch('hasSalaryRange');
   const watchHasExperienceRange = form.watch('hasExperiencerange');
 
-  const [comboBoxInputValue, setComboBoxInputValue] = useState('');
   const [comboBoxSelectedValues, setComboBoxSelectedValues] = useState<
     string[]
   >([]);
-  const [comboBoxLoading, setComboBoxLoading] = useState(false);
-  const [skillDropdownValues, setSkillDropdownValues] = useState<
-    {
-      value: string;
-      label: string;
-    }[]
-  >([{ value: '', label: '' }]);
-
-  const debouncedComboboxValues = useDebounce(comboBoxInputValue, 300);
-
-  async function fetchDropdownValues(searchTerm: string) {
-    setComboBoxLoading(true);
-    fetch(
-      `https://emsiservices.com/skills/versions/latest/skills?q=${searchTerm}&fields=name&limit=5`,
-      {
-        headers: {
-          Authorization: `Bearer ${await getBearerToken()}`,
-        },
-      }
-    )
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-        return response.json();
-      })
-      .then((data: { data: { name: string }[] }) => {
-        const dropdownValues = data.data.map((d) => {
-          return { value: d.name, label: d.name };
-        });
-        setSkillDropdownValues(dropdownValues);
-      })
-      .catch((error) => console.error('Error:', error))
-      .finally(() => {
-        setComboBoxLoading(false);
-      });
-  }
-
-  React.useEffect(() => {
-    !!debouncedComboboxValues.length
-      ? fetchDropdownValues(debouncedComboboxValues)
-      : fetchDropdownValues('javascript');
-  }, [debouncedComboboxValues]);
-
-  React.useEffect(() => {
-    form.setValue('skills', comboBoxSelectedValues);
-  }, [comboBoxSelectedValues, form]);
 
   React.useEffect(() => {
     if (!watchHasSalaryRange) {
@@ -566,53 +516,11 @@ const PostJobForm = () => {
                   </FormItem>
                 )}
               />
-
-              <div className="flex flex-col gap-2">
-                <FormLabel className="font-medium">Skills Required</FormLabel>
-                <Combobox
-                  comboBoxSelectedValues={comboBoxSelectedValues}
-                  setComboBoxSelectedValues={setComboBoxSelectedValues}
-                  setComboBoxInputValue={setComboBoxInputValue}
-                  dropdownValues={skillDropdownValues}
-                  isLoading={comboBoxLoading}
-                ></Combobox>
-              </div>
-
-              <div className="flex flex-wrap gap-2 space-y-0">
-                {comboBoxSelectedValues.map((item, index) => (
-                  <div key={index} className="flex items-csele space-y-0 group">
-                    <div
-                      className={`font-medium text-xs cursor-pointer flex gap-1 text-center justify-start py-2 px-4 rounded-full borderpr-1 bg-blue-100 dark:bg-blue-500 dark:bg-opacity-10 bg-opacity-90 text-blue-700 dark:text-blue-400 border-blue-800 dark:border-blue-400 '}`}
-                    >
-                      {_.startCase(item.toLowerCase())}
-                      {
-                        <Button
-                          className="p-0 h-fit bg-tranparent"
-                          onClick={() => {
-                            setComboBoxSelectedValues((prev) => {
-                              const foundIndex = prev.findIndex(
-                                (val) => val === item
-                              );
-                              if (foundIndex >= 0) {
-                                const updatedComboBoxSelectedValues = [...prev];
-                                updatedComboBoxSelectedValues.splice(
-                                  foundIndex,
-                                  1
-                                );
-                                return updatedComboBoxSelectedValues;
-                              }
-                              return prev;
-                            });
-                          }}
-                        >
-                          <X className="text-white" size={15} />
-                        </Button>
-                      }
-                    </div>
-                  </div>
-                ))}
-                <FormMessage />
-              </div>
+              <SkillsCombobox
+                comboBoxSelectedValues={comboBoxSelectedValues}
+                setComboBoxSelectedValues={setComboBoxSelectedValues}
+                form={form}
+              ></SkillsCombobox>
             </div>
             <div className="bg-gray-900 w-full p-6 rounded-lg space-y-4 mx-auto my-6">
               <h2 className="text-sm text-white capitalize">Job description</h2>
