+++ src/lib/start-cron-job.ts
@@ -1,0 +1,11 @@
+import { startAuthTokenCronJob } from '@/actions/skills.cron';
+
+let isCronJobRunning = false;
+
+function startCronjobWithFlag() {
+  if (!isCronJobRunning) {
+    isCronJobRunning = true;
+    startAuthTokenCronJob();
+  }
+}
+startCronjobWithFlag();
